Update GraphQL implementation for byline fields
This moves all GraphQL functionality to a new file, `inc/graphql.php` and changes the approach for registering GraphQL data so that any `NodeWithAuthor` (post types that support authors) will now have a byline field that will return a list of either `TextProfile` types or `Profile` post types.

To do so, this registers a new object type, called `TextProfile` for manual text profiles, which supports a single text field. This also registers a new union type, `ProfileTypes` that is used to resolve whether a profile is a `TextProfile` or a `Profile` type.