Add block editor support for profiles
This changes the editor used for profile posts to the block editor. This is done by setting `show_in_rest` to `true` when registering the post type. Since this post type is already public and set to use `post` caps, it should not expose any data that isn't already available through other means.